requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
tqdm>=4.65.0
scipy>=1.10.0
scikit-learn>=1.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
transformers>=4.30.0
peft>=0.4.0
bitsandbytes>=0.41.0
accelerate>=0.20.0
pillow>=9.5.0
